slice wrong

+++ src/main.rs
@@ -5,6 +5,7 @@ use std::fs;
 use std::path::Path;
 use std::time::SystemTime;
 use time;
+use std::convert::TryFrom;
 
 #[derive(Debug, Eq, Ord, PartialEq, PartialOrd)]
 struct FileStruct {
@@ -51,7 +52,7 @@ fn list_dir(param: &String) -> std::io::Result<()> {
                         file_result.sort_by_key(|f| f.created_date);
                         let max_file: usize = 5;
                         if total_file > max_file as i32 {
-                            slice_files(file_result, max_file);
+                            slice_files(file_result, max_file, total_file);
                         }
                     }
                 }
@@ -61,22 +62,24 @@ fn list_dir(param: &String) -> std::io::Result<()> {
     Ok(())
 }
 
-fn slice_files(files: Vec<FileStruct>, number: usize) {
-    let total_file = files.len();
-    println!("Length: {:?}", total_file);
-    let files_local = &files[0..files.len()-number];
-    // let files_local = &files[1..files.len()];
-    for file in files_local {
-        println!(
-            "{:?} Date: {:?}",
-            file.name,
-            system_time_to_date_time(file.created_date)
-        );
-        if let Ok(_) = fs::remove_file(file.name.to_string()) {
-            println!(" was delete.");
+fn slice_files(files: Vec<FileStruct>, number: usize, total_file: i32) {
+    if total_file>0 {
+        println!("Length: {:?}", total_file);
+        let delte_total_item = usize::try_from(total_file).unwrap()-number;
+        let files_local = &files[0..delte_total_item];
+        // let files_local = &files[1..files.len()];
+        for file in files_local {
+            println!(
+                "{:?} Date: {:?}",
+                file.name,
+                system_time_to_date_time(file.created_date)
+            );
+            if let Ok(_) = fs::remove_file(file.name.to_string()) {
+                println!(" was delete.");
+            }
         }
+        println!("--------------------------------------------------------------");
     }
-    println!("--------------------------------------------------------------");
 }
 
 fn system_time_to_date_time(t: SystemTime) -> String {
